Add moderator review flag for synced events and review tool
- Set needsModeratorReview on new events imported via external sync
- Add Firestore rules so staff can clear the review flag
- Extend AdminEventsService with needsModeratorReview filter and clear method
- Extract shared events overview widgets from admin events screen
- Add moderator Event Review tool at /moderator/events with review checklist
